🔨 Refactor script for Desmodder
