+++ index.ts
@@ -1,0 +1,68 @@
+	import { Calc, jquery } from "desmodder";
+	import Controller from "./Controller";
+	import { wolfram2desmos } from "./wolfram2desmos"
+
+	let controller: Controller = new Controller(
+		['textarea', 'input'],
+		function (e:FocusEvent) {
+			let elem:HTMLElement|null|undefined = (
+				e.target as HTMLElement
+			)?.parentElement?.parentElement;
+			switch (e.type) {
+				case 'focusin':
+					elem?.addEventListener('paste', pasteHandler, false);
+					break;
+				case 'focusout':
+					elem?.removeEventListener('paste', pasteHandler, false);
+					break;
+				default:
+					break;
+			}
+		}
+	);
+	
+	function typeInTextArea(
+		newText: string | undefined,
+		elm: null | Element = document.activeElement
+	) {
+		const el = elm as HTMLTextAreaElement;
+		const start = el.selectionStart;
+		const end = el.selectionEnd;
+		const text = el.value;
+		const before = text.substring(0, start);
+		const after  = text.substring(end, text.length);
+		el.value = (before + newText + after);
+		el.selectionStart = el.selectionEnd = start + (newText?.length ?? 0);
+		el.focus();
+	}
+	
+	function pasteHandler(e:ClipboardEvent) {
+		let pasteData =  e.clipboardData?.getData('Text');
+		
+		if (pasteData !== '' && Calc.controller.getItemModel(Calc.selectedExpressionId).type == "expression") {
+			// stop data from actually being pasted
+			e.stopPropagation();
+			e.preventDefault();
+			
+			pasteData = wolfram2desmos(pasteData);
+			typeInTextArea(pasteData);
+			
+		}
+	}
+
+	export function onEnable() {
+	  controller.enable();
+	}
+
+	export function onDisable() {
+		controller.disable();
+	}
+
+	export default {
+		id: "wolfram2desmos",
+		name: "Wolfram To Desmos",
+		description: "Convert ASCIImath into Desmos LaTeX on paste.",
+		onEnable: onEnable,
+		onDisable: onDisable,
+		enabledByDefault: true,
+	} as const;
